Fix get_inode to take into account 1-based references
It turns out that in ext2, all the inode references are
done as if tables indexes were 1-based, instead of 0-based.
Because of this, we need to use ino - 1 when getting the inode
from the table.

Additionally, this means the ROOT_INODE is 2 instead of 1. The
very first inode in ext2 is empty for *reasons*.